requirements.txt
python>=3.10
python-docx>=1.1.0
pdfplumber>=0.11.0
openpyxl>=3.1.0
pyyaml>=6.0
pydantic>=2.0
pytest>=7.0
pytest-cov>=4.0
pytest-mock>=3.10
